ui/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "streamlit_example"
version = "0.1.0"
description = "Amazon Bedrock Agentcore User Interface Example for AI Builders Day on 2025/12/20"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.3.2",
    "boto3>=1.40.25",
    "dotenv>=0.9.9",
    "markdown-it-py>=4.0.0",
    "requests>=2.32.5",
    "sseclient-py>=1.8.0",
    "streamlit>=1.49.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
